Delete php/Userstats.php (duplicate file issue)
Oops - Because there were 2 files pushed to dev with the same name "userStats", it was causing weird issues with one file being overwritten by the other each time we merged to dev. On macOS VSCode does not even allow you to create a new file if a file exists with the same name, and casing doesn't change this. I believe the issue is windows allows this, and whoever pushed this from a windows machine didn't realize it was an issue. We received permission from Riad to delete one of the files and update it to the newest file to sort this out other unsuccessful "bug fix" attempts with new task branches (this is because a new branch would have both files causing the same weird issues to happen all over again).